Separate Intra and Outline packages (#70)
* Move intra.go to android/tun2socks.go

* Separate Intra and Outline packages

This slightly reduces the AAR size for the Intra and Outline Android
builds (-1% for Intra, -0.2% for Outline).  More importantly, it
provides a clearer separation of functionality into package trees.

* Avoid repetitive names
