+++ outline/android/tun2socks.go
@@ -19,6 +19,7 @@ import (
 	"math"
 	"runtime/debug"
 
+	"github.com/Jigsaw-Code/outline-go-tun2socks/outline"
 	"github.com/Jigsaw-Code/outline-go-tun2socks/tunnel"
 	"github.com/eycorsican/go-tun2socks/common/log"
 )
@@ -31,7 +32,7 @@ func init() {
 
 // OutlineTunnel embeds the tun2socks.OutlineTunnel interface so it gets exported by gobind.
 type OutlineTunnel interface {
-	tunnel.OutlineTunnel
+	outline.Tunnel
 }
 
 // ConnectShadowsocksTunnel reads packets from a TUN device and routes it to a Shadowsocks proxy server.
@@ -57,7 +58,7 @@ func ConnectShadowsocksTunnel(fd int, host string, port int, password, cipher st
 	if err != nil {
 		return nil, err
 	}
-	t, err := tunnel.NewOutlineTunnel(host, port, password, cipher, isUDPEnabled, tun)
+	t, err := outline.NewTunnel(host, port, password, cipher, isUDPEnabled, tun)
 	if err != nil {
 		return nil, err
 	}
