+++ outline/apple/tun2socks.go
@@ -22,12 +22,12 @@ import (
 	"runtime/debug"
 	"time"
 
-	"github.com/Jigsaw-Code/outline-go-tun2socks/tunnel"
+	"github.com/Jigsaw-Code/outline-go-tun2socks/outline"
 )
 
 // OutlineTunnel embeds the tun2socks.Tunnel interface so it gets exported by gobind.
 type OutlineTunnel interface {
-	tunnel.OutlineTunnel
+	outline.Tunnel
 }
 
 // TunWriter is an interface that allows for outputting packets to the TUN (VPN).
@@ -66,5 +66,5 @@ func ConnectShadowsocksTunnel(tunWriter TunWriter, host string, port int, passwo
 	} else if port <= 0 || port > math.MaxUint16 {
 		return nil, fmt.Errorf("Invalid port number: %v", port)
 	}
-	return tunnel.NewOutlineTunnel(host, port, password, cipher, isUDPEnabled, tunWriter)
+	return outline.NewTunnel(host, port, password, cipher, isUDPEnabled, tunWriter)
 }
